Refactor: Reuses getShiftedTimeRange for both Explore and Dashboards
